requirements.txt
openai
colorama
rich
python-dotenv
langchain==0.3.23
langchain-openai==0.3.12
langgraph==0.2.67
typing-extensions